refactor: Extracted consumer's onreceive message handler function and placed it into the consumer
